+++ transport/transport-runtime/src/main/java/com/google/android/datatransport/runtime/EventInternal.java
@@ -43,6 +43,17 @@ public abstract class EventInternal {
   @Nullable
   public abstract Integer getProductId();
 
+  @Nullable
+  public abstract String getPseudonymousId();
+
+  @Nullable
+  @SuppressWarnings("mutable")
+  public abstract byte[] getExperimentIdsClear();
+
+  @Nullable
+  @SuppressWarnings("mutable")
+  public abstract byte[] getExperimentIdsEncrypted();
+
   public final Map<String, String> getMetadata() {
     return Collections.unmodifiableMap(getAutoMetadata());
   }
@@ -72,6 +83,9 @@ public abstract class EventInternal {
         .setTransportName(getTransportName())
         .setCode(getCode())
         .setProductId(getProductId())
+        .setPseudonymousId(getPseudonymousId())
+        .setExperimentIdsClear(getExperimentIdsClear())
+        .setExperimentIdsEncrypted(getExperimentIdsEncrypted())
         .setEncodedPayload(getEncodedPayload())
         .setEventMillis(getEventMillis())
         .setUptimeMillis(getUptimeMillis())
@@ -97,6 +111,12 @@ public abstract class EventInternal {
     protected abstract Builder setAutoMetadata(Map<String, String> metadata);
 
     public abstract Builder setProductId(Integer value);
+
+    public abstract Builder setPseudonymousId(String value);
+
+    public abstract Builder setExperimentIdsClear(byte[] value);
+
+    public abstract Builder setExperimentIdsEncrypted(byte[] value);
 
     protected abstract Map<String, String> getAutoMetadata();