+++ firebase-sessions/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * [fixed] Handled datastore writes when device has full internal memory more gracefully.
   (GitHub [#5859](https://github.com/firebase/firebase-android-sdk/issues/5859))
-* [fixed] Safely unbind to release service connection.
+* [fixed] Safely unbind malfunctioned session lifecycle service to release service connections.
   (GitHub [#5869](https://github.com/firebase/firebase-android-sdk/issues/5869))
 
 # 1.2.3