+++ firebase-sessions/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Unreleased
 
+* [fixed] Handled datastore writes when device has full internal memory more gracefully.
+  (GitHub [#5859](https://github.com/firebase/firebase-android-sdk/issues/5859))
+* [fixed] Safely unbind malfunctioning session lifecycle service to release service connections.
+  (GitHub [#5869](https://github.com/firebase/firebase-android-sdk/issues/5869))
+
+# 1.2.3
+
+* [fixed] Force validation or rotation of FIDs.
+
+# 1.2.1
+
+* [changed] Bump internal dependencies.
 * [fixed] Handle corruption in DataStore Preferences more gracefully.
 
 # 1.2.0